requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
yt-dlp>=2023.0.0
beautifulsoup4>=4.12.0
selenium>=4.0.0
requests>=2.31.0
transformers>=4.30.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pytest>=7.0.0
markdown>=3.4.0
python-dotenv>=1.0.0
langchain>=0.0.230
torch>=2.0.0
openai>=0.27.0
pillow>=10.0.0
matplotlib>=3.7.0
nltk>=3.8.0
scikit-learn>=1.0.0
opencv-python==4.8.0.74
